Moved routes out of __init__.py
